+++ src/routes/RequestsRoute.js
@@ -111,6 +111,8 @@ import {
   getStatusQuery,
   getFullNameForCsvRecords,
   updateQuerySortString,
+  getPrintedDetails,
+  getLastPrintedDetails,
 } from './utils';
 import SinglePrintButtonForPickSlip from '../components/SinglePrintButtonForPickSlip';
 
@@ -133,15 +135,6 @@ export const getFilteredColumnHeadersMap = (columnHeaders) => (
 
 export const extractPickSlipRequestIds = (pickSlipsData) => {
   return [...new Set(pickSlipsData.map(pickSlip => pickSlip['request.requestID']))];
-};
-
-export const getLastPrintedDetails = (printDetails, intl) => {
-  const fullName = getFullName(printDetails?.lastPrintRequester);
-  const formattedDate = intl.formatDate(printDetails?.printEventDate);
-  const formattedTime = intl.formatTime(printDetails?.printEventDate);
-  const localizedDateTime = `${formattedDate}${formattedTime ? ', ' : ''}${formattedTime}`;
-
-  return fullName + ' ' + localizedDateTime;
 };
 
 export const urls = {
@@ -1040,11 +1033,7 @@ class RequestsRoute extends React.Component {
         record.requester.name = getFullNameForCsvRecords(record.requester);
       }
       if (record.printDetails) {
-        const fullName = getFullNameForCsvRecords(record.printDetails.lastPrintRequester);
-        const lastPrintedDate = record.printDetails.printEventDate || '';
-        const date = lastPrintedDate ? `, ${lastPrintedDate}` : '';
-
-        record.printDetails.lastPrintedDetails = `${fullName}${date}`;
+        record.printDetails.lastPrintedDetails = getPrintedDetails(record);
       }
       if (record.loan) {
         const { dueDate } = record.loan;