+++ src/routes/utils.test.js
@@ -1,3 +1,4 @@
+import { createIntl, createIntlCache } from 'react-intl';
 import {
   requestStatuses,
   requestTypesMap,
@@ -10,6 +11,8 @@ import {
   getStatusQuery,
   getFullNameForCsvRecords,
   updateQuerySortString,
+  getPrintedDetails,
+  getLastPrintedDetails,
 } from './utils';
 
 describe('utils', () => {
@@ -185,6 +188,10 @@ describe('utils', () => {
       const record = {};
       expect(getFullNameForCsvRecords(record)).toBe('');
     });
+
+    it('should return an empty string if the record is missing', () => {
+      expect(getFullNameForCsvRecords()).toBe('');
+    });
   });
 
   describe('updateQuerySortString', () => {
@@ -204,6 +211,66 @@ describe('utils', () => {
       const prop = 'printed,copies';
       const expectedOutput = 'requestDate';
       expect(updateQuerySortString(prop)).toBe(expectedOutput);
+    });
+  });
+
+  describe('getPrintedDetails', () => {
+    it('should return full name when printEventDate is not passed', () => {
+      const record = {
+        printDetails: {
+          lastPrintRequester: { firstName: 'John', lastName: 'Doe' },
+        },
+      };
+
+      expect(getPrintedDetails(record)).toBe('John Doe');
+    });
+
+    it('should return printEventDate when when there is no full name', () => {
+      const record = {
+        printDetails: {
+          lastPrintRequester: {},
+          printEventDate: 'date1',
+        },
+      };
+
+      expect(getPrintedDetails(record)).toBe('date1');
+    });
+
+    it('should return formatted string with full name and last printed date', () => {
+      const record = {
+        printDetails: {
+          lastPrintRequester: { firstName: 'John', lastName: 'Doe' },
+          printEventDate: 'date1',
+        },
+      };
+
+      expect(getPrintedDetails(record)).toBe('John Doe, date1');
+    });
+  });
+
+  describe('getLastPrintedDetails', () => {
+    const intlCache = createIntlCache();
+    const intl = createIntl(
+      {
+        locale: 'en-US',
+        messages: {},
+      },
+      intlCache
+    );
+
+    const lastPrintDetails = {
+      lastPrintRequester: { firstName: 'firstName', middleName: 'middleName', lastName: 'lastName' },
+      lastPrintedDate: '2024-08-03T13:33:31.868Z',
+    };
+
+    it('should return the formatted full name and date/time correctly', () => {
+      const printedDetails = getLastPrintedDetails(lastPrintDetails, intl);
+      const expectedFormattedDate = intl.formatDate(lastPrintDetails.printEventDate);
+      const expectedFormattedTime = intl.formatTime(lastPrintDetails.printEventDate);
+      const expectedOutput =
+        `lastName, firstName middleName ${expectedFormattedDate}${expectedFormattedTime ? ', ' : ''}${expectedFormattedTime}`;
+
+      expect(printedDetails).toBe(expectedOutput);
     });
   });
 });